added role validate and junit test cases

+++ EmployeeDataAPI/src/main/java/com/abn/emsdata/model/Role.java
@@ -1,4 +1,4 @@
 package com.abn.emsdata.model;
 
-public record Role(int id, String name) {
+public record Role(Long id, String name) {
 }
